Configure publishing after project evaluation
This change modifies the `maven-publish` plugin configuration to ensure that the `from(components["release"])` line, which specifies the components to be published, is executed after the project has been fully evaluated. This is done to prevent potential issues where the "release" component might not be available at the time of configuration.